VG-16377 handle transaction with no recipient
Some transaction don't have any recipient. For instance, using remix
IDE you can trigger a transaction to deploy a smart contract, it will
generate a transaction without a `to` field.

This is the quickest solution, perhaps not the cleanest. We just build
an `EthereumTransaction` with an empty string in the `recipient` field
and let the consumer of the live app handle this case.

The other solution would be to change wallet-api types..etc.., but not
sure it is worth it.

Tested on the vault with no issue, not sure what are the immpacts for
ledger-live though.